refactor: 기본 오더 차트 설정에서 적어도 1개 미만으로 지울 수 없게 설정
